Moved SessionInfoHttpHeaders to kapua-job-engine-client module
