+++ job-engine/app/core/pom.xml
@@ -31,7 +31,7 @@
         </dependency>
         <dependency>
             <groupId>org.eclipse.kapua</groupId>
-            <artifactId>kapua-job-engine-jbatch</artifactId>
+            <artifactId>kapua-job-engine-client</artifactId>
         </dependency>
         <dependency>
             <groupId>org.eclipse.kapua</groupId>
