+++ job-engine/app/web/src/main/resources/locator.xml
@@ -88,17 +88,11 @@
 
         <api>org.eclipse.kapua.service.scheduler.trigger.definition.TriggerDefinitionService</api>
         <api>org.eclipse.kapua.service.scheduler.trigger.definition.TriggerDefinitionFactory</api>
-        <!--
-                                        <api>org.eclipse.kapua.service.storable.model.id.StorableIdFactory</api>
-                                        <api>org.eclipse.kapua.service.storable.model.query.predicate.StorablePredicateFactory</api>
--->
 
         <api>org.eclipse.kapua.message.KapuaMessageFactory</api>
         <api>org.eclipse.kapua.message.device.data.KapuaDataMessageFactory</api>
 
         <api>org.eclipse.kapua.transport.TransportClientFactory</api>
-
-        <!--        <api>org.eclipse.kapua.service.config.ServiceConfigurationFactory</api>-->
 
         <api>org.eclipse.kapua.model.id.KapuaIdFactory</api>
         <api>org.eclipse.kapua.model.config.metatype.KapuaMetatypeFactory</api>
