+++ job-engine/client/src/main/java/org/eclipse/kapua/job/engine/client/filter/SessionInfoFilter.java
@@ -10,12 +10,11 @@
  * Contributors:
  *     Eurotech - initial API and implementation
  *******************************************************************************/
-package org.eclipse.kapua.job.engine.client;
+package org.eclipse.kapua.job.engine.client.filter;
 
 import com.google.common.net.HttpHeaders;
 import org.eclipse.kapua.commons.security.KapuaSecurityUtils;
 import org.eclipse.kapua.commons.security.KapuaSession;
-import org.eclipse.kapua.job.engine.JobEngineHttpHeaders;
 import org.eclipse.kapua.job.engine.client.settings.JobEngineClientSetting;
 import org.eclipse.kapua.job.engine.client.settings.JobEngineClientSettingKeys;
 
@@ -24,7 +23,7 @@ import javax.ws.rs.client.ClientRequestFilter;
 import java.io.IOException;
 
 /**
- * {@link ClientRequestFilter} used to populate HTTP request headers with informations about the current {@link KapuaSession}.
+ * {@link ClientRequestFilter} used to populate HTTP request headers with information about the current {@link KapuaSession}.
  *
  * @since 1.5.0
  */
@@ -36,14 +35,14 @@ public class SessionInfoFilter implements ClientRequestFilter {
         String authMode = JobEngineClientSetting.getInstance().getString(JobEngineClientSettingKeys.JOB_ENGINE_CLIENT_AUTH_MODE, "access_token");
         switch (authMode) {
             case "trusted":
-                requestContext.getHeaders().putSingle(JobEngineHttpHeaders.SCOPE_ID_HTTP_HEADER, kapuaSession.getScopeId().toCompactId());
-                requestContext.getHeaders().putSingle(JobEngineHttpHeaders.USER_ID_HTTP_HEADER, kapuaSession.getUserId().toCompactId());
+                requestContext.getHeaders().putSingle(SessionInfoHttpHeaders.SCOPE_ID_HTTP_HEADER, kapuaSession.getScopeId().toCompactId());
+                requestContext.getHeaders().putSingle(SessionInfoHttpHeaders.USER_ID_HTTP_HEADER, kapuaSession.getUserId().toCompactId());
                 break;
             case "access_token":
             default:
                 requestContext.getHeaders().putSingle(HttpHeaders.AUTHORIZATION, String.format("Bearer %s", kapuaSession.getAccessToken().getTokenId()));
         }
-        requestContext.getHeaders().putSingle(JobEngineHttpHeaders.AUTH_MODE, authMode);
+        requestContext.getHeaders().putSingle(SessionInfoHttpHeaders.AUTH_MODE, authMode);
     }
 
 }
